Change folder structure: Move submenu class from helper to its own folder as it is not really a helper in the rails sense
